create BaseService instead of axios/intercept
