+++ src/containers/Polllist/reducer.js
@@ -1,9 +1,4 @@
 import { createSlice } from "@reduxjs/toolkit";
-import axios from "axios";
-import { getPolllistData } from "../../components/Utils";
-import { API_POLL } from "../../constants/api";
-import intercept from "../../axios/interceptors";
-import { REQUEST_STATUS } from "../../constants/status";
 
 const initialState = {
   query: 0,
@@ -43,16 +38,5 @@ export const {
   deletePollSuccess,
 } = slice.actions;
 
-// export const deletePollRequest = () => (dispatch, getState) => {
-//   let pollId = getState().polllist.pollId;
-//   intercept();
-//   axios
-//     .delete(`${API_POLL}/${pollId}`)
-//     .then((respon) => {
-//       if (respon.status === 200) {
-//         console.log("Delete success");
-//       }
-//     })
-//     .catch((err) => console.log(err));
-// };
+
 export default slice.reducer;
