+++ src/service/LoginService.js
@@ -1,6 +1,6 @@
 import { API_SIGNIN } from "../constants/api";
-import axios from "axios";
+import baseService from "./BaseService";
 
 export function login(data) {
-  return axios.post(API_SIGNIN, data);
+  return baseService.post(API_SIGNIN, data);
 }
