+++ src/service/PollDetailService.js
@@ -1,10 +1,10 @@
-import axios from "axios";
 import { API_POLL } from "../constants/api";
+import baseService from "./BaseService";
 
 export function fetchDataPollDetail(id) {
-  return axios.get(`${API_POLL}/${id}`);
+  return baseService.get(`${API_POLL}/${id}`);
 }
 
 export function updatePoll({ pollId, dataUpdate }) {
-  return axios.put(`${API_POLL}/${pollId}`, dataUpdate);
+  return baseService.put(`${API_POLL}/${pollId}`, dataUpdate);
 }
