+++ src/service/PolllistService.js
@@ -1,12 +1,12 @@
-import axios from "axios";
 import { API_POLL } from "../constants/api";
+import baseService from "./BaseService";
 
 export function fetchDataPolllist(query) {
-  return axios.get(
+  return baseService.get(
     `${API_POLL}?offset=${query}&limit=10&direction=desc&search=`
   );
 }
 
 export function deletePollRequest(pollId) {
-  return axios.delete(`${API_POLL}/${pollId}`);
+  return baseService.delete(`${API_POLL}/${pollId}`);
 }
